pyproject.toml
[project]
name = "instaforce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.1",
    "bs4>=0.0.2",
    "chromadb>=1.3.4",
    "duckdb>=1.4.2",
    "duckduckgo-search==5.3.1b1",
    "faiss-cpu>=1.12.0",
    "huggingface-hub>=0.36.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "mysql-connector-python>=9.5.0",
    "numexpr>=2.14.1",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "pytube>=15.0.0",
    "sentence-transformers>=5.1.2",
    "simple-salesforce>=1.12.9",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
    "unstructured>=0.18.20",
    "validators==0.28.1",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.3",
]
